게시글 삭제, 수정 코드 코멘트 반영 수정
