+++ src/main/java/com/uswproject/global/BaseEntity.java
@@ -25,4 +25,8 @@ public abstract class BaseEntity {
 
     @Column(name = "is_deleted", nullable = false)
     private boolean deleted = false;
+
+    public void softDelete() {
+        this.deleted = true;
+    }
 }
